Move mapping/pose_graph/ to internal/.
