+++ cartographer/mapping/internal/2d/pose_graph/spa_cost_function_2d.h
@@ -22,8 +22,8 @@
 #include "Eigen/Core"
 #include "Eigen/Geometry"
 #include "cartographer/common/math.h"
+#include "cartographer/mapping/internal/pose_graph/cost_helpers.h"
 #include "cartographer/mapping/pose_graph.h"
-#include "cartographer/mapping/pose_graph/cost_helpers.h"
 #include "cartographer/transform/rigid_transform.h"
 #include "cartographer/transform/transform.h"
 #include "ceres/ceres.h"
