+++ cartographer/mapping/internal/3d/pose_graph/optimization_problem_3d.cc
@@ -36,7 +36,7 @@
 #include "cartographer/mapping/internal/3d/pose_graph/spa_cost_function_3d.h"
 #include "cartographer/mapping/internal/3d/rotation_cost_function_3d.h"
 #include "cartographer/mapping/internal/3d/rotation_parameterization.h"
-#include "cartographer/mapping/pose_graph/ceres_pose.h"
+#include "cartographer/mapping/internal/pose_graph/ceres_pose.h"
 #include "cartographer/transform/timestamped_transform.h"
 #include "cartographer/transform/transform.h"
 #include "ceres/ceres.h"
