+++ cartographer/mapping/internal/pose_graph/constraint_builder.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cartographer/mapping/pose_graph/constraint_builder.h"
+#include "cartographer/mapping/internal/pose_graph/constraint_builder.h"
 
 #include "cartographer/mapping/2d/scan_matching/ceres_scan_matcher_2d.h"
 #include "cartographer/mapping/2d/scan_matching/fast_correlative_scan_matcher_2d.h"
